backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
python-dotenv
celery
redis
pytest
httpx
